Remove 'dir' parameter to mkstemp
Specifying . as the directory to make the tmpfile in, can cause an exception if the program is run from a directory in which the user does not have write permissions.  For example, if the user does "cd /" and then runs a program which uses this library, wget.download will die with an exception.  I removed the 'dir' parameter which allows mkstemp to choose an appropriate tmp directory for the file.